Let players input their names before play
Add the differentiated method to start a new game based on the given
input names, different from the one that just restarts the game.

Logic to add and remove event listeners decoupled into those related
to buttons and those related to the game board, so the button ones
don't get removed at the end of each game.